requirements.txt
google-adk
kiteconnect
pandas
numpy
plotly
gemini-api
